consumer/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
kafka-python==2.2.15
multidict==6.6.3
packaging==25.0
pact-python==2.3.3
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
